[Refactor] Disentangle HWCustomOp and HLSBackend parts of attention ops
