+++ src/finn/custom_op/fpgadataflow/hls/__init__.py
@@ -27,6 +27,11 @@
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 from finn.custom_op.fpgadataflow.hls.addstreams_hls import AddStreams_hls
+from finn.custom_op.fpgadataflow.hls.attention_heads_hls import (
+    MergeMultiHeads_hls,
+    SplitMultiHeads_hls,
+)
+from finn.custom_op.fpgadataflow.hls.attention_hls import ScaledDotProductAttention_hls
 from finn.custom_op.fpgadataflow.hls.channelwise_op_hls import ChannelwiseOp_hls
 from finn.custom_op.fpgadataflow.hls.checksum_hls import CheckSum_hls
 from finn.custom_op.fpgadataflow.hls.concat_hls import StreamingConcat_hls
@@ -43,6 +48,7 @@ from finn.custom_op.fpgadataflow.hls.labelselect_hls import LabelSelect_hls
 from finn.custom_op.fpgadataflow.hls.lookup_hls import Lookup_hls
 from finn.custom_op.fpgadataflow.hls.matrixvectoractivation_hls import MVAU_hls
 from finn.custom_op.fpgadataflow.hls.pool_hls import Pool_hls
+from finn.custom_op.fpgadataflow.hls.replicate_stream_hls import ReplicateStream_hls
 from finn.custom_op.fpgadataflow.hls.streamingdatawidthconverter_hls import (
     StreamingDataWidthConverter_hls,
 )
@@ -80,18 +86,7 @@ custom_op["UpsampleNearestNeighbour_hls"] = UpsampleNearestNeighbour_hls
 custom_op["MVAU_hls"] = MVAU_hls
 custom_op["VVAU_hls"] = VVAU_hls
 
-from finn.custom_op.fpgadataflow.attention import (
-    ScaledDotProductAttention as ScaledDotProductAttention_hls
-)
-from finn.custom_op.fpgadataflow.attention_heads import (
-    SplitMultiHeads as SplitMultiHeads_hls,
-    MergeMultiHeads as MergeMultiHeads_hls
-)
-from finn.custom_op.fpgadataflow.replicate_stream import (
-    ReplicateStream as ReplicateStream_hls
-)
-
-custom_op["ScaledDotProductAttention_hls"]  = ScaledDotProductAttention_hls
+custom_op["ScaledDotProductAttention_hls"] = ScaledDotProductAttention_hls
 custom_op["SplitMultiHeads_hls"] = SplitMultiHeads_hls
 custom_op["MergeMultiHeads_hls"] = MergeMultiHeads_hls
 custom_op["ReplicateStream_hls"] = ReplicateStream_hls
